[PERFECTIVE] Added the visualization of address spaces (unfinished)
git-svn-id: svn://svn.code.sf.net/p/kactus2/code/trunk@205 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ common/utils.h
@@ -167,9 +167,9 @@ namespace Utils {
 	 * 
 	 * \param str The string to convert.
 	 *
-	 * \return int The result of the conversion.
+	 * \return The result of the conversion.
 	*/
-	int str2Int(const QString& str);
+	quint64 str2Int(const QString& str);
 }
 
 #endif // UTILS_H
